src/requirements.txt
fastapi
uvicorn[standard]
gunicorn
alembic
SQLAlchemy
psycopg2-binary
python-dotenv
python-multipart
pydantic
pydantic-settings
passlib
python-jose
email-validator